feat: add a portable flock(1)
To be able to port some nice Linux utility, one first need a working
flock.

There is a repo for that: https://github.com/discoteq/flock but the
current homebrew install is broken because it is too old.

This version specifically download the last commit (at the time of
creation) and package it.